requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
python-dotenv==1.0.0
beautifulsoup4==4.12.2
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2 